requirements.txt
llama-index-core>=0.10.0
llama-index-vector-stores-elasticsearch>=0.1.0
llama-index-embeddings-openai>=0.1.0
elasticsearch>=8.0.0
python-dotenv>=1.0.0
openai>=1.0.0